refactor TokenClient/Extensions
